RAG Chatbot/requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
langchain==0.1.0
google-generativeai>=0.4.0  # Compatible with langchain-google-genai 2.0.10
faiss-cpu==1.7.4
python-dotenv==1.0.0
langchain-google-genai==2.0.10
langchain-community==0.0.28
pydantic==2.5.0